requirements.txt
boto3==1.9.16
botocore==1.12.16
docutils==0.14
jmespath==0.9.3
pika==0.12.0
python-dateutil==2.7.3
PyYAML==3.13
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.23
